# public/api-docs/mailchimp_openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "MailchimpAPIGateway",
    "description" : "Mailchimp API Gateway",
    "contact" : { },
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8000/mailchimp"
  } ],
  "paths" : {
...
    "/lists/{listId}/members" : {
      "get" : {
        "tags" : [ "Members" ],
        "summary" : "Show members",
        "description" : "Return all members of a list",
        "operationId" : "ShowMembers",
        "parameters" : [ {
          "name" : "listId",
          "in" : "path",
          "description" : "list ID under which member is to be created",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "post" : {
        "tags" : [ "Members" ],
        "summary" : "Create member",
        "description" : "Return a member",
        "operationId" : "CreateMember",
        "parameters" : [ {
          "name" : "listId",
          "in" : "path",
          "description" : "list ID under which member is to be created",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/body"
              }
            }
          },
          "required" : false
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/lists/{listId}/members/{memberId}" : {
      "get" : {
        "tags" : [ "Members" ],
        "summary" : "Show member",
        "description" : "Return a specific member by ID",
        "operationId" : "ShowMember",
        "parameters" : [ {
          "name" : "listId",
          "in" : "path",
          "description" : "list ID under which member is to get",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "memberId",
          "in" : "path",
          "description" : "ID of the member to get",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "put" : {
        "tags" : [ "Members" ],
        "summary" : "Update member",
        "description" : "Return the specific updated member",
        "operationId" : "UpdateMember",
        "parameters" : [ {
          "name" : "Content-Type",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "example" : "application/json"
          }
        }, {
          "name" : "listId",
          "in" : "path",
          "description" : "list ID under which member is to be updated",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "memberId",
          "in" : "path",
          "description" : "ID of the member to be updated",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdatememberRequest"
              },
              "example" : "{\"email_address\":\"RoyaltyXCoLtdX@hotmail.com\",\"status\":\"subscribed\",\"language\":\"USX EnglishX\",\"vip\":true,\"location\":{\"latitude\":\"-37.898725\",\"longitude\":\"145.049333\"},\"ip_signup\":\"172.198.34.87\",\"tags\":[\"Soccer\",\"Fashion\"]}"
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "delete" : {
        "tags" : [ "Members" ],
        "summary" : "Delete member",
        "description" : "Delete a member by ID. Return nothing",
        "operationId" : "DeleteMember",
        "parameters" : [ {
          "name" : "listId",
          "in" : "path",
          "description" : "list ID under which member is to be deleted",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "memberId",
          "in" : "path",
          "description" : "ID of the member to be deleted",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "UpdatememberRequest" : {
        "title" : "UpdatememberRequest",
        "required" : [ "email_address", "ip_signup", "language", "location", "status", "tags", "vip" ],
        "type" : "object",
        "properties" : {
          "email_address" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "language" : {
            "type" : "string"
          },
          "vip" : {
            "type" : "boolean"
          },
          "location" : {
            "$ref" : "#/components/schemas/Location"
          },
          "ip_signup" : {
            "type" : "string"
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "example" : {
          "email_address" : "RoyaltyXCoLtdX@hotmail.com",
          "status" : "subscribed",
          "language" : "USX EnglishX",
          "vip" : true,
          "location" : {
            "latitude" : "-37.898725",
            "longitude" : "145.049333"
          },
          "ip_signup" : "172.198.34.87",
          "tags" : [ "Soccer", "Fashion" ]
        }
      },
      "Location" : {
        "title" : "Location",
        "required" : [ "latitude", "longitude" ],
        "type" : "object",
        "properties" : {
          "latitude" : {
            "type" : "string"
          },
          "longitude" : {
            "type" : "string"
          }
        }
      },
      "body" : {
        "required" : [ "email_address", "ip_signup", "language", "location[latitude]", "location[longitude]", "status", "tags[]", "vip" ],
        "type" : "object",
        "properties" : {
          "email_address" : {
            "type" : "string",
            "example" : "RoyaltyCoLtd@hotmail.com"
          },
          "status" : {
            "type" : "string",
            "example" : "subscribed"
          },
          "language" : {
            "type" : "string",
            "example" : "US English"
          },
          "vip" : {
            "type" : "string",
            "example" : "1"
          },
          "location[latitude]" : {
            "type" : "string",
            "example" : "-37.898725"
          },
          "location[longitude]" : {
            "type" : "string",
            "example" : "145.049333"
          },
          "ip_signup" : {
            "type" : "string",
            "example" : "172.198.34.87"
          },
          "tags[]" : {
            "type" : "string",
            "example" : "Soccer"
          }
        }
      }
    }
  }
}